refactor(core): add @o3co/auth-provider-core/testing subpath + ADR augmentations
Completes the schema-strict refactor (Option C) by exposing the shared
test-fixture factory through a public testing subpath, addressing the
five cross-package fixture sites broken by 94317fc8 and the ADR
augmentations / minor findings raised by the plaas-thread
/multi-agent-review run.

Changes:
- Move src/__tests__/fixtures/valid-config.mts to src/testing/fixtures/
  and re-export through src/testing/index.mts. Add the ./testing
  subpath to packages/core/package.json#exports, mapped to
  dist/testing/index.mjs (+.d.mts). Per A2-gamma spec section 6.1 and
  section 7.
- Update twelve internal core test imports to use the new path; remove
  the old __tests__/fixtures/ directory.
- Replace inline minimal-config construction in five cross-package
  boot integration tests (redis x3, federation-google, federation-github)
  with imports from @o3co/auth-provider-core/testing. The bare
  { http: {}, oauth: { jwt: {}, ... } } shapes used previously can no
  longer parse against the strict schema; the testing subpath provides
  a shared baseline so each package no longer reinvents the minimal
  shape.
- ADR augmentations:
  - I2 (library-consumer perspective): consumers embedding the package
    outside the standalone HOCON path are responsible for supplying
    defaults before validate(); the testing factories provide a
    reference baseline.
  - I4 (federations.<name>.enabled strict): operators must now declare
    enabled = true / false explicitly or omit the federation entry;
    the implicit-disable behaviour is gone. Documents the breaking
    change.
  - P2 (fixture honesty): docstring on the moved fixture now describes
    it as "minimal schema-valid config" rather than claiming to mirror
    application.conf, since the values intentionally diverge for test
    ergonomics.
  - M3 (path consistency): "Pattern preserved" list now uses full paths
    (oauth.jwt.signingKey.local.*, session.storage.redis.password)
    instead of unqualified suffixes.
- CHANGELOG.md adds the testing subpath under Unreleased > Added and a
  Breaking Changes entry covering the schema-strict refactor with the
  federations.enabled and library-consumer migration notes.
- Minor polish (M1 / M2 from plaas-thread review): config.test.mts test
  names rephrased from "defaults to X" to "is X when application.conf
  is loaded with no override" so the reader sees the value comes from
  hocon, not the schema; comment beside the algorithm/kid assertion
  cites ADR 2026-04-30 explicitly. validate-manifests.test.mts step-13
  test renames its module-specific config field from port to
  cfgModRetries to remove the visual collision with http.port from
  CoreConfigSchema.

Verification:
- make test (auth.provider workspace, including redis testcontainers):
  1889 tests pass across all packages.
- pnpm -r run typecheck: clean across the six packages with a typecheck
  script.
- node --input-type=module -e "import { makeValidAppConfig } from
  '@o3co/auth-provider-core/testing'": the subpath resolves from
  packages/redis and the factories return the expected shape.

Carries forward 94317fc8 unchanged. Lands as PR alpha; PR beta
(roadmap Phase 9-11) follows once this is merged.
