+++ packages/core/package.json
@@ -14,6 +14,10 @@
 		"./modules/manifest": {
 			"import": "./dist/modules/manifest/index.mjs",
 			"types": "./dist/modules/manifest/index.d.mts"
+		},
+		"./testing": {
+			"import": "./dist/testing/index.mjs",
+			"types": "./dist/testing/index.d.mts"
 		}
 	},
 	"files": [
