+++ packages/core/src/boot/__tests__/create-app.test.mts
@@ -28,7 +28,7 @@
  */
 
 import { describe, expect, it } from "vitest";
-import { makeValidCoreConfig } from "../../__tests__/fixtures/valid-config.mjs";
+import { makeValidCoreConfig } from "../../testing/fixtures/valid-config.mjs";
 import { defineModule } from "../../modules/manifest/index.mjs";
 import { createApp } from "../create-app.mjs";
 import type {
