+++ packages/core/src/boot/__tests__/integration.test.mts
@@ -29,7 +29,7 @@
 
 import { Router } from "express";
 import { describe, expect, it } from "vitest";
-import { makeValidCoreConfig } from "../../__tests__/fixtures/valid-config.mjs";
+import { makeValidCoreConfig } from "../../testing/fixtures/valid-config.mjs";
 import { createBootApp } from "../../index.mjs";
 import { defineModule } from "../../modules/manifest/index.mjs";
 import type { BootstrapMap } from "../types.mjs";
