+++ packages/core/src/boot/__tests__/materialize-components.test.mts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 import { describe, expect, it, vi } from "vitest";
-import { makeValidCoreConfig } from "../../__tests__/fixtures/valid-config.mjs";
+import { makeValidCoreConfig } from "../../testing/fixtures/valid-config.mjs";
 import { defineModule } from "../../modules/manifest/index.mjs";
 import { materializeComponents } from "../materialize-components.mjs";
 import { planBoot } from "../plan-boot.mjs";
