+++ packages/redis/__tests__/redisSessionStoresModule.test.mts
@@ -15,6 +15,7 @@
  */
 
 import { BootError, createBootApp, defineModule } from "@o3co/auth-provider-core";
+import { makeValidCoreConfig } from "@o3co/auth-provider-core/testing";
 import Redis from "ioredis";
 import { GenericContainer, type StartedTestContainer } from "testcontainers";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
@@ -36,8 +37,7 @@ afterAll(async () => {
 
 const minBoot = (extra: Record<string, unknown>) =>
 	({
-		http: {},
-		oauth: { jwt: {}, accessToken: {}, refreshToken: {}, grants: {} },
+		...makeValidCoreConfig(),
 		...extra,
 	}) as never;
